move Province enum to province.py
