+++ policyengine_canada/variables/household/demographic/geographic/province/province.py
@@ -1,7 +1,21 @@
 from policyengine_canada.model_api import *
-from policyengine_canada.variables.household.demographic.geographic.province.province_enum import (
-    Province,
-)
+
+
+class Province(Enum):
+    UNKNOWN = "Unknown"
+    ALBERTA = "Alberta"
+    BRITISH_COLUMBIA = "British Columbia"
+    MANITOBA = "Manitoba"
+    NEW_BRUNSWICK = "New Brunswick"
+    NEWFOUNDLAND_AND_LABRADOR = "Newfoundland and Labrador"
+    NOVA_SCOTIA = "Nova Scotia"
+    NORTHWEST_TERRITORIES = "Northwest Territories"
+    NUNAVUT = "Nunavut"
+    ONTARIO = "Ontario"
+    PRINCE_EDWARD_ISLAND = "Prince Edward Island"
+    QUEBEC = "Quebec"
+    SASKATCHEWAN = "Saskatchewan"
+    YUKON = "Yukon"
 
 
 class province(Variable):
