--- policyengine_canada/variables/household/demographic/geographic/province/province_enum.py
@@ -1,18 +1,0 @@
-from policyengine_canada.model_api import *
-
-
-class Province(Enum):
-    UNKNOWN = "Unknown"
-    ALBERTA = "Alberta"
-    BRITISH_COLUMBIA = "British Columbia"
-    MANITOBA = "Manitoba"
-    NEW_BRUNSWICK = "New Brunswick"
-    NEWFOUNDLAND_AND_LABRADOR = "Newfoundland and Labrador"
-    NOVA_SCOTIA = "Nova Scotia"
-    NORTHWEST_TERRITORIES = "Northwest Territories"
-    NUNAVUT = "Nunavut"
-    ONTARIO = "Ontario"
-    PRINCE_EDWARD_ISLAND = "Prince Edward Island"
-    QUEBEC = "Quebec"
-    SASKATCHEWAN = "Saskatchewan"
-    YUKON = "Yukon"
